requirements.txt
PyAutoGUI
python-rtmidi
pyyaml
